avoid reset of UTILITY_TIMER
resetting the UTILITY_TIMER makes accurate long term time measurement
much harder. We never need to reset it if we treat it as a wrapping 16
bit unsigned number and always do timing maths with 16 bit subtraction

+++ Mcu/f051/Src/stm32f0xx_it.c
@@ -46,7 +46,7 @@ extern char out_put;
 extern char compute_dshot_flag;
 /* USER CODE END EV */
 
-int interrupt_time = 0;
+uint16_t interrupt_time = 0;
 /******************************************************************************/
 /*           Cortex-M0 Processor Interruption and Exception Handlers */
 /******************************************************************************/
